+++ CHANGELOG.md
@@ -9,6 +9,10 @@ Changes to prior versions can be found on the [GitHub release page](https://gith
 
 ## [Unreleased](https://github.com/cryptomator/cryptofs/compare/2.10.0...HEAD)
 
+### Added
+* Property `vaultId` in `CryptoFileSystemProperties` to set the vault id during vault initialization ([#358](https://github.com/cryptomator/cryptofs/pull/358))
+* Getter `UnverifiedVaultConfig::allegedVaultId` to get the vault id of an unverified vault config ([#358](https://github.com/cryptomator/cryptofs/pull/358))
+
 ### Changed
 * Updated dependencies ([#357](https://github.com/cryptomator/cryptofs/pull/357)):
   * `com.auth0:java-jwt` from 4.5.1 to 4.6.1